+++ src/components/guider/GuiderStats.vue
@@ -1,8 +1,8 @@
 <template>
   <p class="font-bold">{{ $t('components.guider.rms_error') }}:</p>
-  <p>RA: {{ store.guiderInfo.RMSError.RA.Arcseconds.toFixed(2) }}</p>
-  <p>DEC: {{ store.guiderInfo.RMSError.Dec.Arcseconds.toFixed(2) }}</p>
-  <p>Total: {{ store.guiderInfo.RMSError.Total.Arcseconds.toFixed(2) }}</p>
+  <p>RA: {{ store.guiderInfo?.RMSError?.RA.Arcseconds.toFixed(2) }}</p>
+  <p>DEC: {{ store.guiderInfo.RMSError?.Dec.Arcseconds.toFixed(2) }}</p>
+  <p>Total: {{ store.guiderInfo.RMSError?.Total.Arcseconds.toFixed(2) }}</p>
 </template>
 
 <script setup>