+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.0.9.0] - 
+### Added
+- Guidgraph can be displayed everywhere. It can be opened and closed from the status bar
+- Add skychart to sequenz
+
+### Changed
+- Connection timeout increased to 5s
+
+### Fixed
+- Automatic reconnect of the mount websocket connection
 
 ## [1.0.8.0] - 2025-05-27
 ### Added